+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/metadata/ShardSpecMixIn.java
@@ -15,6 +15,7 @@ import io.druid.timeline.partition.NoneShardSpec;
         @JsonSubTypes.Type(value = NumberedShardSpec.class, name = "hashed"),
         @JsonSubTypes.Type(value = NumberedShardSpec.class, name = "linear"),
         @JsonSubTypes.Type(value = NumberedShardSpec.class, name = "numbered"),
+        @JsonSubTypes.Type(value = IdOnlyShardSpec.class, name = "single"),
         @JsonSubTypes.Type(value = NoneShardSpec.class, name = "none")
 })
 public abstract class ShardSpecMixIn { }